+++ tests/PersistentModelTest.php
@@ -2,6 +2,7 @@
 
 namespace Rubix\ML\Tests;
 
+use Rubix\ML\Verbose;
 use Rubix\ML\Ranking;
 use Rubix\ML\Learner;
 use Rubix\ML\Wrapper;
@@ -42,6 +43,7 @@ class PersistentModelTest extends TestCase
         $this->assertInstanceOf(Wrapper::class, $this->estimator);
         $this->assertInstanceOf(Probabilistic::class, $this->estimator);
         $this->assertInstanceOf(Ranking::class, $this->estimator);
+        $this->assertInstanceOf(Verbose::class, $this->estimator);
         $this->assertInstanceOf(Learner::class, $this->estimator);
         $this->assertInstanceOf(Estimator::class, $this->estimator);
     }
@@ -68,7 +70,7 @@ class PersistentModelTest extends TestCase
     public function params() : void
     {
         $expected = [
-            'estimator' => new DummyClassifier(),
+            'base' => new DummyClassifier(),
             'persister' => new Filesystem('test.model'),
         ];